requirements.txt
flask
python-dotenv
line-bot-sdk
pytesseract
Pillow
gunicorn
